pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ti"
version = "0.1.0"
description = "Ti — Topological Investigation of the Riemann Hypothesis via Sheaf-Theoretic Gauge Fields"
requires-python = ">=3.11"
dependencies = [
    "numpy>=1.24",
    "scipy>=1.11",
    "h5py>=3.9",
    "matplotlib>=3.7",
]

[project.optional-dependencies]
gpu = ["cupy-cuda12x"]
dev = ["pytest>=7.4", "pytest-cov>=4.1", "scikit-learn>=1.3"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
